// src/Types/types.ts
export enum FilterStatus {
  All = 'all',
  Active = 'active',
  Completed = 'completed',
}

export enum Actions {
  SetTodos = 'SET_TODOS',
  AddTodo = 'ADD_TODO',
  UpdateTodo = 'UPDATE_TODO',
  DeleteTodo = 'DELETE_TODO',
  ToggleAllTodos = 'TOGGLE_ALL_TODOS',
  ClearCompleted = 'CLEAR_COMPLETED',
}